+++ include/usb_osintf.h
@@ -22,7 +22,7 @@
 
 #include <usb_vendor_req.h>
 
-#define USBD_HALTED(Status) ((ULONG)(Status) >> 30 == 3)
+#define USBD_HALTED(Status) ((u32)(Status) >> 30 == 3)
 
 
 //uint usb_dvobj_init(struct rtl_priv * rtlpriv);
